requirements.txt
google-generativeai
python-dotenv
flask